SS trying to refactor calendar
