+++ src/calendar.rs
@@ -15,9 +15,19 @@ use std::path::Path;
 use std::process::Command;
 use std::str::FromStr;
 
-// Remove unused constants
-// const ALL_DAY_DURATION: i64 = 86400;
-// const DEFAULT_DURATION: i64 = 3600;
+mod calendar_types;
+mod calendar_validation;
+mod calendar_applescript;
+mod calendar_import;
+mod calendar_contacts;
+#[cfg(test)]
+mod calendar_tests;
+
+pub use calendar_types::*;
+pub use calendar_validation::*;
+pub use calendar_applescript::*;
+pub use calendar_import::*;
+pub use calendar_contacts::*;
 
 /// Custom error type for calendar operations
 #[derive(Debug, thiserror::Error)]
@@ -34,345 +44,6 @@ pub enum CalendarError {
 
     #[error("AppleScript execution failed: {0}")]
     ScriptError(String),
-}
-
-/// Recurrence frequency for repeating events
-#[derive(Debug, Clone, PartialEq, Copy)]
-pub enum RecurrenceFrequency {
-    Daily,
-    Weekly,
-    Monthly,
-    Yearly,
-}
-
-impl RecurrenceFrequency {
-    #[allow(dead_code)]
-    pub fn to_applescript(&self) -> &'static str {
-        match self {
-            RecurrenceFrequency::Daily => "daily",
-            RecurrenceFrequency::Weekly => "weekly",
-            RecurrenceFrequency::Monthly => "monthly",
-            RecurrenceFrequency::Yearly => "yearly",
-        }
-    }
-
-    /// Convert to RFC 5545 format for iCalendar
-    pub fn to_rfc5545(&self) -> &'static str {
-        match self {
-            RecurrenceFrequency::Daily => "DAILY",
-            RecurrenceFrequency::Weekly => "WEEKLY",
-            RecurrenceFrequency::Monthly => "MONTHLY",
-            RecurrenceFrequency::Yearly => "YEARLY",
-        }
-    }
-
-    /// Parse recurrence frequency from string
-    pub fn from_str(s: &str) -> Result<Self> {
-        match s.to_lowercase().as_str() {
-            "daily" | "day" | "days" => Ok(RecurrenceFrequency::Daily),
-            "weekly" | "week" | "weeks" => Ok(RecurrenceFrequency::Weekly),
-            "monthly" | "month" | "months" => Ok(RecurrenceFrequency::Monthly),
-            "yearly" | "year" | "years" | "annual" | "annually" => Ok(RecurrenceFrequency::Yearly),
-            _ => Err(anyhow!("Invalid recurrence frequency: {}", s)),
-        }
-    }
-}
-
-/// Recurrence pattern for calendar events
-#[derive(Debug, Clone)]
-pub struct RecurrencePattern {
-    /// Frequency of recurrence
-    pub frequency: RecurrenceFrequency,
-
-    /// Interval between occurrences (e.g., every 2 weeks)
-    pub interval: u32,
-
-    /// End date of recurrence (None for indefinite)
-    pub end_date: Option<String>,
-
-    /// Number of occurrences (None for indefinite)
-    pub count: Option<u32>,
-
-    /// Days of the week for weekly recurrence (0=Sunday, 1=Monday, etc.)
-    pub days_of_week: Vec<u8>,
-}
-
-impl RecurrencePattern {
-    /// Create a new simple recurrence pattern with the given frequency
-    pub fn new(frequency: RecurrenceFrequency) -> Self {
-        Self { frequency, interval: 1, end_date: None, count: None, days_of_week: Vec::new() }
-    }
-
-    /// Set the interval for recurrence
-    pub fn with_interval(mut self, interval: u32) -> Self {
-        self.interval = interval;
-        self
-    }
-
-    /// Set the end date for recurrence
-    pub fn with_end_date(mut self, end_date: &str) -> Self {
-        self.end_date = Some(end_date.to_string());
-        self
-    }
-
-    /// Set the count of occurrences
-    pub fn with_count(mut self, count: u32) -> Self {
-        self.count = Some(count);
-        self
-    }
-
-    /// Set the days of week for weekly recurrence
-    pub fn with_days_of_week(mut self, days: &[u8]) -> Self {
-        self.days_of_week = days.to_vec();
-        self
-    }
-}
-
-/// Configuration for a calendar event
-#[derive(Debug, Clone)]
-pub struct EventConfig {
-    pub title: String,
-    pub start_date: String,
-    pub start_time: String,
-    #[allow(dead_code)]
-    pub end_date: Option<String>,
-    pub end_time: Option<String>,
-    pub calendars: Vec<String>,
-    pub all_day: bool,
-    pub location: Option<String>,
-    pub description: Option<String>,
-    pub emails: Vec<String>,
-    pub reminder: Option<i32>,
-    pub timezone: Option<String>,
-    pub recurrence: Option<RecurrencePattern>,
-    // Enhanced Zoom integration fields
-    pub create_zoom_meeting: bool,
-    #[allow(dead_code)]
-    pub zoom_meeting_id: Option<u64>,
-    #[allow(dead_code)]
-    pub zoom_join_url: Option<String>,
-    #[allow(dead_code)]
-    pub zoom_password: Option<String>,
-}
-
-impl EventConfig {
-    pub fn new(title: &str, date: &str, time: &str) -> Self {
-        Self {
-            title: title.to_string(),
-            start_date: date.to_string(),
-            start_time: time.to_string(),
-            end_date: None,
-            end_time: None,
-            calendars: Vec::new(),
-            all_day: false,
-            location: None,
-            description: None,
-            emails: Vec::new(),
-            reminder: None,
-            timezone: None,
-            recurrence: None,
-            // Initialize Zoom fields
-            create_zoom_meeting: false,
-            zoom_meeting_id: None,
-            zoom_join_url: None,
-            zoom_password: None,
-        }
-    }
-
-    // Method to easily set recurrence pattern
-    #[allow(dead_code)]
-    pub fn with_recurrence(mut self, recurrence: RecurrencePattern) -> Self {
-        self.recurrence = Some(recurrence);
-        self
-    }
-
-    // Method to enable Zoom meeting
-    #[allow(dead_code)]
-    pub fn with_zoom_meeting(mut self, enable: bool) -> Self {
-        self.create_zoom_meeting = enable;
-        self
-    }
-
-    pub fn validate(&self) -> Result<()> {
-        // Validate date format (YYYY-MM-DD)
-        if !validate_date_format(&self.start_date) {
-            return Err(CalendarError::InvalidDateTime(format!(
-                "Invalid date format: {}",
-                self.start_date
-            ))
-            .into());
-        }
-
-        // Validate time format (HH:MM)
-        if !validate_time_format(&self.start_time) {
-            return Err(CalendarError::InvalidDateTime(format!(
-                "Invalid time format: {}",
-                self.start_time
-            ))
-            .into());
-        }
-
-        // Validate end time if specified
-        if let Some(end_time) = &self.end_time {
-            if !validate_time_format(end_time) {
-                return Err(CalendarError::InvalidDateTime(format!(
-                    "Invalid end time format: {}",
-                    end_time
-                ))
-                .into());
-            }
-        }
-
-        // Process title to safely handle quotes from NLP-generated commands
-        // First trim quotes if they're at the start and end
-        let mut sanitized_title = self.title.trim_matches('"').to_string();
-
-        // Handle escaped quotes as well
-        if sanitized_title.contains("\\\"") {
-            sanitized_title = sanitized_title.replace("\\\"", "");
-        }
-
-        debug!("Original title: '{}', Sanitized title: '{}'", self.title, sanitized_title);
-
-        // Check for truly dangerous characters (excluding quotes which are handled above)
-        if sanitized_title.contains(';')
-            || sanitized_title.contains('&')
-            || sanitized_title.contains('|')
-            || sanitized_title.contains('<')
-            || sanitized_title.contains('>')
-            || sanitized_title.contains('$')
-        {
-            return Err(anyhow!("Title contains potentially dangerous characters"));
-        }
-
-        // Validate location if specified
-        if let Some(location) = &self.location {
-            let mut sanitized_location = location.replace("\\\"", "").replace("\"", "");
-            if sanitized_location.starts_with('"') && sanitized_location.ends_with('"') {
-                sanitized_location =
-                    sanitized_location[1..sanitized_location.len() - 1].to_string();
-            }
-
-            if contains_dangerous_characters(&sanitized_location) {
-                return Err(anyhow!("Location contains potentially dangerous characters"));
-            }
-        }
-
-        // Validate description if specified
-        if let Some(description) = &self.description {
-            if contains_dangerous_chars_for_script(description) {
-                return Err(anyhow!("Description contains potentially dangerous characters"));
-            }
-        }
-
-        // Validate emails
-        for email in &self.emails {
-            if !validate_email(email) {
-                return Err(anyhow!("Invalid email format: {}", email));
-            }
-        }
-
-        // Validate timezone if specified
-        if let Some(timezone) = &self.timezone {
-            // Basic validation - more comprehensive would check against a list of valid timezones
-            if timezone.len() > 50 || contains_dangerous_chars_for_script(timezone) {
-                return Err(anyhow!("Invalid timezone format"));
-            }
-        }
-
-        // Validate recurrence if specified
-        if let Some(recurrence) = &self.recurrence {
-            if let Some(end_date) = &recurrence.end_date {
-                if !validate_date_format(end_date) {
-                    return Err(anyhow!("Invalid recurrence end date format: {}", end_date));
-                }
-            }
-        }
-
-        // If creating a Zoom meeting, validate needed fields
-        if self.create_zoom_meeting {
-            // Zoom requires an end time for calculating meeting duration
-            if self.end_time.is_none() {
-                return Err(anyhow!("End time is required for Zoom meetings"));
-            }
-        }
-
-        Ok(())
-    }
-}
-
-/// Validate date string has format YYYY-MM-DD
-pub fn validate_date_format(date: &str) -> bool {
-    let re = Regex::new(r"^\d{4}-\d{2}-\d{2}$").unwrap();
-
-    if !re.is_match(date) {
-        return false;
-    }
-
-    // Further validate the date is reasonable
-    if let Ok(naive_date) = chrono::NaiveDate::parse_from_str(date, "%Y-%m-%d") {
-        // Check date is within reasonable range
-        let year = naive_date.year();
-        return year >= 2000 && year <= 2100;
-    }
-
-    false
-}
-
-/// Validate time string has format HH:MM
-pub fn validate_time_format(time: &str) -> bool {
-    let re = Regex::new(r"^\d{1,2}:\d{2}$").unwrap();
-
-    if !re.is_match(time) {
-        return false;
-    }
-
-    // Further validate the time values
-    let parts: Vec<&str> = time.split(':').collect();
-    if parts.len() != 2 {
-        return false;
-    }
-
-    if let (Ok(hours), Ok(minutes)) = (parts[0].parse::<u32>(), parts[1].parse::<u32>()) {
-        return hours < 24 && minutes < 60;
-    }
-
-    false
-}
-
-/// Enhanced email validation to handle edge cases and improve error reporting
-pub fn validate_email(email: &str) -> bool {
-    let re = Regex::new(r"^[A-Za-z0-9._%+-]{1,64}@(?:[A-Za-z0-9-]{1,63}\.){1,125}[A-Za-z]{2,63}$")
-        .unwrap();
-
-    if !re.is_match(email) {
-        debug!("Invalid email format detected: {}", email);
-        return false;
-    }
-
-    // Check for dangerous characters that could cause script injection
-    if contains_dangerous_characters(email) {
-        debug!("Email contains dangerous characters: {}", email);
-        return false;
-    }
-
-    true
-}
-
-/// Check for potentially dangerous characters that could cause AppleScript injection
-fn contains_dangerous_characters(input: &str) -> bool {
-    // Note: Double quotes are handled separately in title/location sanitization
-    input.contains(';')
-        || input.contains('&')
-        || input.contains('|')
-        || input.contains('<')
-        || input.contains('>')
-        || input.contains('$')
-}
-
-/// Check for characters that could break AppleScript specifically
-fn contains_dangerous_chars_for_script(input: &str) -> bool {
-    input.contains('\"') || input.contains('\\') || input.contains('¬')
 }
 
 pub async fn list_calendars() -> Result<()> {
@@ -425,8 +96,11 @@ pub async fn list_calendars() -> Result<()> {
 pub async fn create_event(config: EventConfig) -> Result<()> {
     debug!("Creating event with config: {:?}", config);
 
+    // Fix: Bring validate into scope for EventConfig
+    use crate::calendar::calendar_validation::validate_event_config;
+
     // Validate the event configuration first
-    config.validate()?;
+    validate_event_config(&config)?;
 
     // First verify Calendar.app is running and get available calendars
     ensure_calendar_running().await?;
@@ -437,7 +111,7 @@ pub async fn create_event(config: EventConfig) -> Result<()> {
 
     // Load configuration and get default calendar if none specified
     let app_config = Config::load()?;
-    let requested_calendars = if (config.calendars.is_empty()) {
+    let requested_calendars = if config.calendars.is_empty() {
         vec![app_config.calendar.default_calendar.unwrap_or_else(|| "Calendar".to_string())]
     } else {
         // Validate that specified calendars exist
@@ -907,111 +581,14 @@ async fn ensure_calendar_running() -> Result<()> {
     if output.status.success() { Ok(()) } else { Err(CalendarError::NotRunning.into()) }
 }
 
-pub async fn list_event_properties() -> Result<()> {
-    ensure_calendar_running().await?;
-
-    let script = r#"tell application "Calendar"
-        try
-            set propList to {}
-            
-            -- Basic properties that we can set
-            copy "summary (title)" to end of propList
-            copy "start date" to end of propList
-            copy "end date" to end of propList
-            copy "allday" to end of propList
-            copy "description" to end of propList
-            copy "location" to end of propList
-            copy "url" to end of propList
-            copy "calendar" to end of propList
-            copy "recurrence" to end of propList
-            copy "status" to end of propList
-            copy "availability" to end of propList
-            
-            return propList
-        on error errMsg
-            error "Failed to get event properties: " & errMsg
-        end try
-    end tell"#;
-
-    let output = tokio::process::Command::new("osascript").arg("-e").arg(script).output().await?;
-
-    if output.status.success() {
-        println!("Available Calendar Event Properties:");
-        let properties = String::from_utf8_lossy(&output.stdout);
-        if !properties.trim().is_empty() {
-            for prop in properties.trim_matches('{').trim_matches('}').split(", ") {
-                println!("  - {}", prop.trim_matches('"'));
-            }
-        } else {
-            println!("  No properties found. Calendar might not be accessible.");
-        }
-        Ok(())
-    } else {
-        Err(anyhow!(
-            "Failed to get event properties: {}",
-            String::from_utf8_lossy(&output.stderr)
-        ))
-    }
-}
-
-#[allow(dead_code)]
-pub async fn delete_event(title: &str, _date: &str) -> Result<()> {
-    let script = format!(
-        r#"tell application "Calendar"
-            try
-                set foundEvents to {{}}
-                repeat with c in calendars
-                    tell c
-                        set matchingEvents to (every event whose summary contains "{0}")
-                        repeat with evt in matchingEvents
-                            copy evt to end of foundEvents
-                        end repeat
-                    end tell
-                end repeat
-                
-                if (count of foundEvents) is 0 then
-                    error "No matching events found for title: {0}"
-                end if
-                
-                repeat with evt in foundEvents
-                    set evtTitle to summary of evt
-                    log "Deleting event: " & evtTitle
-                    delete evt
-                end repeat
-                
-                return "Success: Events deleted"
-            on error errMsg
-                return "Error: " & errMsg
-            end try
-        end tell"#,
-        title
-    );
-
-    let output = tokio::process::Command::new("osascript")
-        .arg("-e")
-        .arg(&script)
-        .output()
-        .await?;
-
-    let result = String::from_utf8_lossy(&output.stdout);
-    if result.contains("Success") {
-        println!("Calendar event(s) deleted containing title: {}", title);
-        Ok(())
-    } else {
-        Err(anyhow!("Failed to delete events: {}", result))
-    }
-}
-
 /// Lookup a contact by name and return their email addresses
 pub async fn lookup_contact(name: &str) -> Result<Vec<String>> {
-    debug!("Looking up contact by name: '{}'", name);
-
-    // Try exact match first
-    let exact_script = format!(
+    debug!("Looking up contact: {}", name);
+    let script = format!(
         r#"tell application "Contacts"
             set the_emails to {{}}
             try
-                set the_people to (every person whose name is "{}")
+                set the_people to (every person whose name contains "{}")
                 repeat with the_person in the_people
                     if exists email of the_person then
                         repeat with the_email in (get every email of the_person)
@@ -1032,632 +609,35 @@ pub async fn lookup_contact(name: &str) -> Result<Vec<String>> {
 
     let output = tokio::process::Command::new("osascript")
         .arg("-e")
-        .arg(&exact_script)
+        .arg(&script)
         .output()
         .await
         .map_err(|e| anyhow!("Failed to execute AppleScript: {}", e))?;
 
     if output.status.success() {
         let emails = String::from_utf8_lossy(&output.stdout);
-        debug!("Raw contact lookup output (exact match): '{}'", emails);
+        debug!("Raw contact lookup output: {}", emails);
         let email_list: Vec<String> = emails
             .trim_matches('{')
             .trim_matches('}')
             .split(", ")
             .filter(|s| !s.is_empty() && !s.contains("missing value"))
             .map(|s| s.trim_matches('"').trim().to_string())
-            .filter(|email| validate_email(email))
             .collect();
-
-        if !email_list.is_empty() {
-            info!("Found {} email(s) for contact '{}': {:?}", email_list.len(), name, email_list);
-            return Ok(email_list);
-        }
-    }
-
-    // If exact match returned no results, try contains match
-    debug!("No exact match found for '{}', trying contains match", name);
-    let contains_script = format!(
-        r#"tell application "Contacts"
-            set the_emails to {{}}
-            try
-                set searchText to "{}"
-                set the_people to (every person whose name contains searchText)
-                repeat with the_person in the_people
-                    set personName to name of the_person
-                    if personName contains searchText then
-                        if exists email of the_person then
-                            repeat with the_email in (get every email of the_person)
-                                if value of the_email is not missing value then
-                                    set the end of the_emails to (value of the_email as text)
-                                end if
-                            end repeat
-                        end if
-                    end if
-                end repeat
-                return the_emails
-            on error errMsg
-                log "Error looking up contact: " & errMsg
-                return {{}}
-            end try
-        end tell"#,
-        name.replace("\"", "\\\"")
-    );
-
-    let output = tokio::process::Command::new("osascript")
-        .arg("-e")
-        .arg(&contains_script)
-        .output()
-        .await
-        .map_err(|e| anyhow!("Failed to execute AppleScript: {}", e))?;
-
-    if output.status.success() {
-        let emails = String::from_utf8_lossy(&output.stdout);
-        debug!("Raw contact lookup output (contains match): '{}'", emails);
-        let email_list: Vec<String> = emails
-            .trim_matches('{')
-            .trim_matches('}')
-            .split(", ")
-            .filter(|s| !s.is_empty() && !s.contains("missing value"))
-            .map(|s| s.trim_matches('"').trim().to_string())
-            .filter(|email| validate_email(email))
-            .collect();
-
-        if !email_list.is_empty() {
-            info!(
-                "Found {} email(s) for '{}' with contains match: {:?}",
+        if email_list.is_empty() {
+            debug!("No emails found for contact '{}'", name);
+        } else {
+            debug!(
+                "Found {} email(s) for '{}': {:?}",
                 email_list.len(),
                 name,
                 email_list
             );
-        } else {
-            debug!("No emails found for contact '{}'", name);
         }
-
-        return Ok(email_list);
+        Ok(email_list)
     } else {
         let error = String::from_utf8_lossy(&output.stderr);
         error!("Contact lookup error: {}", error);
         Ok(Vec::new())
-    }
-}
-
-/// Enhanced event creation with contact lookup
-pub async fn create_event_with_contacts(config: EventConfig, contact_names: &[&str]) -> Result<()> {
-    info!("Creating event with {} contact names: {:?}", contact_names.len(), contact_names);
-    let mut found_emails = Vec::new();
-
-    for name in contact_names {
-        info!("Looking up contact: '{}'", name);
-        match lookup_contact(name).await {
-            Ok(emails) => {
-                if emails.is_empty() {
-                    info!("No email found for contact: '{}'", name);
-                } else {
-                    info!("Found {} email(s) for contact '{}': {:?}", emails.len(), name, emails);
-                    found_emails.extend(emails.into_iter().filter(|e| validate_email(e)));
-                }
-            }
-            Err(e) => {
-                error!("Failed to lookup contact '{}': {}", name, e);
-            }
-        }
-    }
-
-    let mut config = EventConfig {
-        emails: Vec::with_capacity(config.emails.len() + found_emails.len()),
-        ..config
-    };
-
-    // First add any existing emails
-    if !config.emails.is_empty() {
-        info!("Adding {} existing email(s): {:?}", config.emails.len(), config.emails);
-        config.emails.extend(config.emails.clone());
-    }
-
-    // Then add newly found emails
-    if !found_emails.is_empty() {
-        info!("Adding {} found email(s): {:?}", found_emails.len(), found_emails);
-        config.emails.extend(found_emails);
-
-        config.emails.sort_unstable();
-        config.emails.dedup();
-        info!("Final email list after deduplication: {}", config.emails.len());
-    } else {
-        info!("No emails found for any of the contacts");
-    }
-
-    create_event(config).await
-}
-
-pub async fn import_csv_events(file_path: &Path, target_calendar: Option<String>) -> Result<()> {
-    let file = File::open(file_path)?;
-    let mut rdr = ReaderBuilder::new()
-        .delimiter(b',') // Use comma as delimiter for CSV
-        .has_headers(true)
-        .flexible(true)
-        .trim(csv::Trim::All)
-        .comment(Some(b'#'))
-        .from_reader(file);
-
-    // Read headers
-    let headers: Vec<String> = rdr.headers()?.iter().map(|h| h.trim().to_lowercase()).collect();
-
-    debug!("Found headers: {:?}", headers);
-
-    // Validate required headers are present
-    let required_headers = ["title", "date", "start_time"];
-    for header in &required_headers {
-        if !headers.contains(&header.to_string()) {
-            return Err(anyhow!(
-                "Required header '{}' not found in CSV. Required headers: title, date, start_time",
-                header
-            ));
-        }
-    }
-
-    let header_positions: std::collections::HashMap<String, usize> =
-        headers.iter().enumerate().map(|(i, h)| (h.to_string(), i)).collect();
-
-    let mut success_count = 0;
-    let mut error_count = 0;
-
-    // Get default calendar for fallback
-    let app_config = Config::load()?;
-    let default_calendar =
-        app_config.calendar.default_calendar.unwrap_or_else(|| "Calendar".to_string());
-
-    // Process each record
-    for (row_num, result) in rdr.records().enumerate() {
-        let record = match result {
-            Ok(r) => r,
-            Err(e) => {
-                error!("Failed to read row {}: {}", row_num + 1, e);
-                error_count += 1;
-                continue;
-            }
-        };
-        debug!("Processing record values: {:?}", record);
-
-        // Extract required fields with safe access
-        let title = match record.get(header_positions["title"]) {
-            Some(t) if !t.trim().is_empty() => t.trim(),
-            _ => {
-                error!("Row {}: Missing or empty title", row_num + 1);
-                error_count += 1;
-                continue;
-            }
-        };
-
-        let date = match record.get(header_positions["date"]) {
-            Some(d) if validate_date_format(d.trim()) => d.trim(),
-            _ => {
-                error!(
-                    "Row {}: Invalid or missing date format (required: YYYY-MM-DD)",
-                    row_num + 1
-                );
-                error_count += 1;
-                continue;
-            }
-        };
-
-        let start_time = match record.get(header_positions["start_time"]) {
-            Some(t) if validate_time_format(t.trim()) => t.trim(),
-            _ => {
-                error!("Row {}: Invalid or missing time format (required: HH:MM)", row_num + 1);
-                error_count += 1;
-                continue;
-            }
-        };
-
-        let mut config = EventConfig::new(title, date, start_time);
-
-        // Set optional end time if present
-        if let Some(&pos) = header_positions.get("end_time") {
-            if let Some(end_time) = record.get(pos) {
-                let end_time = end_time.trim();
-                if !end_time.is_empty() {
-                    if validate_time_format(end_time) {
-                        config.end_time = Some(end_time.to_string());
-                    } else {
-                        error!("Row {}: Invalid end time format (required: HH:MM)", row_num + 1);
-                        error_count += 1;
-                        continue;
-                    }
-                }
-            }
-        }
-
-        // Collect attendee emails
-        let mut attendee_emails = Vec::new();
-
-        // Handle calendar field - if it's an email, add as attendee
-        if let Some(&pos) = header_positions.get("calendar") {
-            if let Some(value) = record.get(pos) {
-                let value = value.trim();
-                if !value.is_empty() && validate_email(value) {
-                    debug!("Found email in calendar field: {}", value);
-                    attendee_emails.push(value.to_string());
-                }
-            }
-        }
-
-        // Handle attendees field
-        if let Some(&pos) = header_positions.get("attendees") {
-            if let Some(attendees) = record.get(pos) {
-                if !attendees.is_empty() {
-                    debug!("Processing attendees field: {}", attendees);
-                    for email in attendees.split(|c| c == ';' || c == ',') {
-                        let email = email.trim();
-                        if !email.is_empty() && validate_email(email) {
-                            debug!("Found valid email in attendees: {}", email);
-                            attendee_emails.push(email.to_string());
-                        }
-                    }
-                }
-            }
-        }
-
-        // Set calendar - prefer command line target, then default
-        if let Some(cal) = &target_calendar {
-            debug!("Using command line calendar: {}", cal);
-            config.calendars = vec![cal.to_string()];
-        } else {
-            debug!("Using default calendar: {}", default_calendar);
-            config.calendars = vec![default_calendar.clone()];
-        }
-
-        // Set description
-        if let Some(&pos) = header_positions.get("description") {
-            if let Some(desc) = record.get(pos) {
-                let desc = desc.trim();
-                if !desc.is_empty() {
-                    config.description = Some(desc.to_string());
-                }
-            }
-        }
-
-        // Set location
-        if let Some(&pos) = header_positions.get("location") {
-            if let Some(loc) = record.get(pos) {
-                let loc = loc.trim();
-                if !loc.is_empty() {
-                    config.location = Some(loc.to_string());
-                }
-            }
-        }
-
-        // Set deduplicated attendees
-        if !attendee_emails.is_empty() {
-            attendee_emails.sort();
-            attendee_emails.dedup();
-            debug!(
-                "Setting {} attendees for event: {}",
-                attendee_emails.len(),
-                attendee_emails.join(", ")
-            );
-            config.emails = attendee_emails;
-        }
-
-        // Create the event
-        debug!("Creating event with config: {:?}", config);
-        match create_event(config).await {
-            Ok(_) => {
-                success_count += 1;
-            }
-            Err(e) => {
-                error_count += 1;
-                error!("Failed to import event: {}", e);
-            }
-        }
-    }
-
-    if success_count > 0 || error_count > 0 {
-        println!(
-            "Import completed: {} events imported successfully, {} failed",
-            success_count, error_count
-        );
-    } else {
-        println!("No events found in the CSV file");
-    }
-    Ok(())
-}
-
-pub async fn import_ics_events(file_path: &Path, target_calendar: Option<String>) -> Result<()> {
-    let buf = BufReader::new(File::open(file_path)?);
-    let parser = IcalParser::new(buf);
-
-    let mut success_count = 0;
-    let mut error_count = 0;
-
-    for calendar in parser {
-        match calendar {
-            Ok(cal) => {
-                for event in cal.events {
-                    if let Err(e) = import_ical_event(event, &target_calendar).await {
-                        error_count += 1;
-                        error!("Failed to import ICS event: {}", e);
-                    } else {
-                        success_count += 1;
-                    }
-                }
-            }
-            Err(e) => {
-                error!("Failed to parse ICS calendar: {}", e);
-                return Err(anyhow!("Failed to parse ICS file: {}", e));
-            }
-        }
-    }
-
-    println!(
-        "Import completed: {} events imported successfully, {} failed",
-        success_count, error_count
-    );
-    Ok(())
-}
-
-async fn import_ical_event(event: IcalEvent, target_calendar: &Option<String>) -> Result<()> {
-    // Get required properties
-    let summary = event
-        .properties
-        .iter()
-        .find(|p| p.name == "SUMMARY")
-        .and_then(|p| p.value.as_ref())
-        .ok_or_else(|| anyhow!("Event missing SUMMARY"))?;
-
-    let dt_start = event
-        .properties
-        .iter()
-        .find(|p| p.name == "DTSTART")
-        .and_then(|p| p.value.as_ref())
-        .ok_or_else(|| anyhow!("Event missing DTSTART"))?;
-
-    // Parse start datetime
-    let start_dt = NaiveDateTime::parse_from_str(dt_start, "%Y%m%dT%H%M%S")?;
-
-    let mut config = EventConfig::new(
-        summary,
-        &start_dt.format("%Y-%m-%d").to_string(),
-        &start_dt.format("%H:%M").to_string(),
-    );
-
-    // Set target calendar if specified
-    if let Some(cal) = target_calendar {
-        config.calendars = vec![cal.to_string()];
-    }
-
-    // Get optional end time
-    if let Some(dt_end) = event
-        .properties
-        .iter()
-        .find(|p| p.name == "DTEND")
-        .and_then(|p| p.value.as_ref())
-    {
-        if let Ok(end_dt) = NaiveDateTime::parse_from_str(dt_end, "%Y%m%dT%H%M%S") {
-            config.end_time = Some(end_dt.format("%H:%M").to_string());
-        }
-    }
-
-    // Get description
-    if let Some(desc) = event
-        .properties
-        .iter()
-        .find(|p| p.name == "DESCRIPTION")
-        .and_then(|p| p.value.as_ref())
-    {
-        config.description = Some(desc.to_string());
-    }
-
-    // Get location
-    if let Some(loc) = event
-        .properties
-        .iter()
-        .find(|p| p.name == "LOCATION")
-        .and_then(|p| p.value.as_ref())
-    {
-        config.location = Some(loc.to_string());
-    }
-
-    // Get attendees
-    let attendees: Vec<String> = event
-        .properties
-        .iter()
-        .filter(|p| p.name == "ATTENDEE")
-        .filter_map(|p| p.value.as_ref())
-        .map(|v| v.trim_start_matches("mailto:").to_string())
-        .collect();
-
-    if !attendees.is_empty() {
-        config.emails = attendees;
-    }
-
-    // Handle recurrence rule if present
-    if let Some(rrule) = event
-        .properties
-        .iter()
-        .find(|p| p.name == "RRULE")
-        .and_then(|p| p.value.as_ref())
-    {
-        if let Some(recurrence) = parse_ical_recurrence(rrule) {
-            config.recurrence = Some(recurrence);
-        }
-    }
-
-    create_event(config).await
-}
-
-fn parse_ical_recurrence(rrule: &str) -> Option<RecurrencePattern> {
-    let parts = rrule
-        .split(';') // Removed mut as it's not needed
-        .map(|s| s.split('='))
-        .filter_map(|mut kv| Some((kv.next()?, kv.next()?)));
-
-    let mut frequency = None;
-    let mut interval = 1;
-    let mut end_date = None;
-    let mut count = None;
-    let mut days = Vec::new();
-
-    for (key, value) in parts {
-        match key {
-            "FREQ" => {
-                frequency = match value {
-                    "DAILY" => Some(RecurrenceFrequency::Daily),
-                    "WEEKLY" => Some(RecurrenceFrequency::Weekly),
-                    "MONTHLY" => Some(RecurrenceFrequency::Monthly),
-                    "YEARLY" => Some(RecurrenceFrequency::Yearly),
-                    _ => None,
-                };
-            }
-            "INTERVAL" => {
-                if let Ok(val) = value.parse() {
-                    interval = val;
-                }
-            }
-            "UNTIL" => {
-                if let Ok(dt) = NaiveDateTime::parse_from_str(value, "%Y%m%dT%H%M%SZ") {
-                    end_date = Some(dt.format("%Y-%m-%d").to_string());
-                }
-            }
-            "COUNT" => {
-                if let Ok(val) = value.parse() {
-                    count = Some(val);
-                }
-            }
-            "BYDAY" => {
-                days = value
-                    .split(',')
-                    .filter_map(|day| match day {
-                        "SU" => Some(0),
-                        "MO" => Some(1),
-                        "TU" => Some(2),
-                        "WE" => Some(3),
-                        "TH" => Some(4),
-                        "FR" => Some(5),
-                        "SA" => Some(6),
-                        _ => None,
-                    })
-                    .collect();
-            }
-            _ => {}
-        }
-    }
-
-    frequency.map(|f| {
-        let mut pattern = RecurrencePattern::new(f).with_interval(interval);
-        if let Some(end) = end_date {
-            pattern = pattern.with_end_date(&end);
-        }
-        if let Some(c) = count {
-            pattern = pattern.with_count(c);
-        }
-        if !days.is_empty() {
-            pattern = pattern.with_days_of_week(&days);
-        }
-        pattern
-    })
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use chrono::{NaiveDate, Timelike};
-
-    // ...existing code...
-
-    #[test]
-    fn test_event_import_csv_header_validation() {
-        // Test basic CSV header validation requirements
-        let required = ["title", "date", "time"];
-
-        let valid_headers =
-            vec!["title", "date", "time", "location", "description", "attendees", "calendar"];
-        assert!(required.iter().all(|&req| valid_headers.contains(&req)));
-
-        let missing_title = vec!["date", "time", "location", "description"];
-        assert!(!required.iter().all(|&req| missing_title.contains(&req)));
-    }
-
-    #[test]
-    fn test_event_import_csv_date_time_parsing() {
-        // Test parsing of date and time from CSV format
-        let date_str = "2025-03-20";
-        let time_str = "10:30";
-
-        let date = NaiveDate::parse_from_str(date_str, "%Y-%m-%d").unwrap();
-        let time = chrono::NaiveTime::parse_from_str(time_str, "%H:%M").unwrap();
-        let dt = chrono::NaiveDateTime::new(date, time);
-
-        assert_eq!(dt.year(), 2025);
-        assert_eq!(dt.month(), 3);
-        assert_eq!(dt.day(), 20);
-        assert_eq!(dt.hour(), 10);
-        assert_eq!(dt.minute(), 30);
-    }
-
-    #[test]
-    fn test_event_import_email_validation() {
-        // Test email validation for attendees and calendar fields
-        assert!(validate_email("user@example.com"));
-        assert!(validate_email("user.name@example.co.uk"));
-        assert!(!validate_email("invalid-email"));
-    }
-
-    #[test]
-    fn test_csv_event_import_email_validation() {
-        // Test email validation used in CSV import
-        assert!(validate_email("user@example.com"));
-        assert!(validate_email("user.name@example.co.uk"));
-        assert!(!validate_email("invalid-email"));
-        assert!(!validate_email("user@"));
-        assert!(!validate_email("@example.com"));
-    }
-
-    #[test]
-    fn test_csv_date_time_parsing() {
-        // Test CSV date and time format parsing
-        let date = NaiveDate::parse_from_str("2025-03-20", "%Y-%m-%d").unwrap();
-        let time = chrono::NaiveTime::parse_from_str("10:30", "%H:%M").unwrap();
-        let dt = chrono::NaiveDateTime::new(date, time);
-
-        assert_eq!(dt.year(), 2025);
-        assert_eq!(dt.month(), 3);
-        assert_eq!(dt.day(), 20);
-        assert_eq!(dt.hour(), 10);
-        assert_eq!(dt.minute(), 30);
-    }
-
-    #[test]
-    fn test_csv_header_requirements() {
-        // Test CSV header validation requirements
-        let required = ["title", "date", "time"];
-
-        let valid_headers = vec!["title", "date", "time", "location", "description"];
-        assert!(required.iter().all(|&req| valid_headers.contains(&req)));
-
-        let missing_title = vec!["date", "time", "location", "description"];
-        assert!(!required.iter().all(|&req| missing_title.contains(&req)));
-    }
-
-    #[test]
-    fn test_date_parsing() {
-        // Test parsing and components of date strings
-        let date_str = "2025-04-15";
-        let time_str = "10:30";
-
-        let full_dt_str = format!("{} {}", date_str, time_str);
-        let dt = NaiveDateTime::parse_from_str(&full_dt_str, "%Y-%m-%d %H:%M").unwrap();
-
-        // Verify each component is correctly parsed
-        assert_eq!(dt.year(), 2025, "Year should be 2025");
-        assert_eq!(dt.month(), 4, "Month should be 4 (April)");
-        assert_eq!(dt.day(), 15, "Day should be 15");
-        assert_eq!(dt.hour(), 10, "Hour should be 10");
-        assert_eq!(dt.minute(), 30, "Minute should be 30");
-
-        // Test formatting of components for AppleScript
-        assert_eq!(dt.format("%Y").to_string(), "2025");
-        assert_eq!(dt.format("%-m").to_string(), "4"); // Month without leading zero
-        assert_eq!(dt.format("%-d").to_string(), "15"); // Day without leading zero
     }
 }
