+++ src/calendar/calendar_applescript.rs
@@ -1,0 +1,41 @@
+//! AppleScript and Calendar.app integration for DuckTape calendar module.
+//
+// This module provides async functions for interacting with macOS Calendar.app via AppleScript.
+
+use anyhow::Result;
+
+/// Ensure Calendar.app is running
+pub async fn ensure_calendar_running() -> Result<()> {
+    // ...implementation moved from calendar.rs...
+    Ok(())
+}
+
+/// List all calendars
+pub async fn list_calendars() -> Result<()> {
+    // ...implementation moved from calendar.rs...
+    Ok(())
+}
+
+/// Get available calendars
+pub async fn get_available_calendars() -> Result<Vec<String>> {
+    // ...implementation moved from calendar.rs...
+    Ok(vec![])
+}
+
+/// Create a single event in Calendar.app
+pub async fn create_single_event(/* params */) -> Result<()> {
+    // ...implementation moved from calendar.rs...
+    Ok(())
+}
+
+/// List event properties
+pub async fn list_event_properties() -> Result<()> {
+    // ...implementation moved from calendar.rs...
+    Ok(())
+}
+
+/// Delete an event by title and date
+pub async fn delete_event(title: &str, date: &str) -> Result<()> {
+    // ...implementation moved from calendar.rs...
+    Ok(())
+}
