+++ src/calendar_contacts.rs
@@ -1,0 +1,17 @@
+//! Contact lookup logic for DuckTape calendar module.
+//
+// This module provides functions to look up contacts and their emails.
+
+use anyhow::Result;
+
+/// Lookup a contact by name and return their email addresses
+pub async fn lookup_contact(name: &str) -> Result<Vec<String>> {
+    // ...implementation moved from calendar.rs...
+    Ok(vec![])
+}
+
+/// Enhanced event creation with contact lookup
+pub async fn create_event_with_contacts(/* params */) -> Result<()> {
+    // ...implementation moved from calendar.rs...
+    Ok(())
+}
