+++ src/calendar_import.rs
@@ -1,0 +1,30 @@
+//! CSV and ICS import logic for DuckTape calendar module.
+//
+// This module provides functions to import events from CSV and ICS files.
+
+use anyhow::Result;
+use std::path::Path;
+
+/// Import events from a CSV file
+pub async fn import_csv_events(file_path: &Path, target_calendar: Option<String>) -> Result<()> {
+    // ...implementation moved from calendar.rs...
+    Ok(())
+}
+
+/// Import events from an ICS file
+pub async fn import_ics_events(file_path: &Path, target_calendar: Option<String>) -> Result<()> {
+    // ...implementation moved from calendar.rs...
+    Ok(())
+}
+
+/// Import a single iCal event
+pub async fn import_ical_event(/* params */) -> Result<()> {
+    // ...implementation moved from calendar.rs...
+    Ok(())
+}
+
+/// Parse iCal recurrence rule
+pub fn parse_ical_recurrence(rrule: &str) -> Option<crate::calendar_types::RecurrencePattern> {
+    // ...implementation moved from calendar.rs...
+    None
+}
