+++ src/calendar_tests.rs
@@ -1,0 +1,8 @@
+//! Unit tests for the calendar module.
+//
+// This module contains tests for calendar validation, types, and integration logic.
+
+#[cfg(test)]
+mod tests {
+    // ...tests moved from calendar.rs...
+}
